requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
asqlite @ git+https://github.com/Rapptz/asqlite.git@ed54062e35acb5d0fe73bf536fc9dd6f22886950
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
frozenlist==1.3.3
h11==0.14.0
idna==3.4
iso8601==2.0.0
multidict==6.0.4
python-dotenv==1.0.0
sniffio==1.3.0
sse-starlette==1.6.1
starlette==0.28.0
twitchio==2.6.0
typing_extensions==4.7.1
uvicorn==0.22.0
yarl==1.9.2
